Added statuses to sale_items

+++ application/controllers/Admin.php
@@ -126,10 +126,16 @@ class Admin extends MY_Controller
     public function confirm_order($order_id)
     {
         if (_is_user_login($this)) {
+            $user = _get_current_user_id($this);
             $this->load->model("product_model");
             $order = $this->product_model->get_sale_order_by_id($order_id);
-            if (!empty($order)) {
-                $this->db->query("update sale set status = 1 where sale_id = '" . $order_id . "'");
+            $orderItems = $this->product_model->get_sale_order_items($order_id, $user);
+            if (!empty($order) && !empty($orderItems)) {
+                $this->db->query("update sale_items set item_status = 1 where sale_id = '" . $order_id . "'");
+                $orderItems = $this->product_model->get_sale_order_items_by_order_id_with_status($order_id, 0);
+                if (empty($orderItems)) {
+                    $this->db->query("update sale set status = 1 where sale_id = '" . $order_id . "'");
+                }
                 $q = $this->db->query("Select * from registers where user_id = '" . $order->user_id . "'");
                 $user = $q->row();
 
@@ -160,9 +166,12 @@ class Admin extends MY_Controller
     public function delivered_order($order_id)
     {
         if (_is_user_login($this)) {
+            $user = _get_current_user_id($this);
             $this->load->model("product_model");
             $order = $this->product_model->get_sale_order_by_id($order_id);
-            if (!empty($order)) {
+            $orderItems = $this->product_model->get_sale_order_items($order_id, $user);
+            if (!empty($order) && !empty($orderItems)) {
+                $this->db->query("update sale_items set item_status = 2 where sale_id = '" . $order_id . "'");
                 $this->db->query("update sale set status = 2 where sale_id = '" . $order_id . "'");
                 /* $this->db->query("INSERT INTO delivered_order (sale_id, user_id, on_date, delivery_time_from, delivery_time_to, status, note, is_paid, total_amount, total_rewards, total_kg, total_items, socity_id, delivery_address, location_id, delivery_charge, new_store_id, assign_to, payment_method)
                 SELECT sale_id, user_id, on_date, delivery_time_from, delivery_time_to, status, note, is_paid, total_amount, total_rewards, total_kg, total_items, socity_id, delivery_address, location_id, delivery_charge, new_store_id, assign_to, payment_method FROM sale
@@ -201,9 +210,12 @@ class Admin extends MY_Controller
     public function delivered_order_complete($order_id)
     {
         if (_is_user_login($this)) {
+            $user = _get_current_user_id($this);
             $this->load->model("product_model");
             $order = $this->product_model->get_sale_order_by_id($order_id);
-            if (!empty($order)) {
+            $orderItems = $this->product_model->get_sale_order_items($order_id, $user);
+            if (!empty($order) && !empty($orderItems)) {
+                $this->db->query("update sale_items set item_status = 4 where sale_id = '" . $order_id . "'");
                 $this->db->query("update sale set status = 4 where sale_id = '" . $order_id . "'");
                 $this->db->query("INSERT INTO delivered_order (sale_id, user_id, on_date, delivery_time_from, delivery_time_to, status, note, is_paid, total_amount, total_rewards, total_kg, total_items, socity_id, delivery_address, location_id, delivery_charge, new_store_id, assign_to, payment_method)
                 SELECT sale_id, user_id, on_date, delivery_time_from, delivery_time_to, status, note, is_paid, total_amount, total_rewards, total_kg, total_items, socity_id, delivery_address, location_id, delivery_charge, new_store_id, assign_to, payment_method FROM sale
@@ -250,10 +262,12 @@ class Admin extends MY_Controller
     public function cancle_order($order_id)
     {
         if (_is_user_login($this)) {
+            $user = _get_current_user_id($this);
             $this->load->model("product_model");
             $order = $this->product_model->get_sale_order_by_id($order_id);
-
-            if (!empty($order)) {
+            $orderItems = $this->product_model->get_sale_order_items($order_id, $user);
+            if (!empty($order) && !empty($orderItems)) {
+                $this->db->query("update sale_items set item_status = 3 where sale_id = '" . $order_id . "'");
                 $this->db->query("update sale set status = 3 where sale_id = '" . $order_id . "'");
                 $this->db->delete('sale_items', array('sale_id' => $order_id));
